Kotlin Coroutines & Flows Coroutine Basics
[Hard] Assignment #3

Create a single-screen app using coroutines in Compose. Add three buttons, each representing a different bird. Create a composable function for each bird that displays the bird’s name and launches a coroutine to print the bird's sound to the console. Only render the composable of the bird represented by the last button tapped and ensure only the selected bird's sounds are printed.